+++ src/util/data/FlagVal.java
@@ -31,8 +31,14 @@ public class FlagVal {
     }
     public static FlagVal newVal(String combined){
         String[] spl = combined.split("_");
-        if( spl.length==2)
+        if( spl.length==2) {
             return new FlagVal().group(spl[0]).name(spl[1]);
+        }else if( spl.length>2){
+            String name = spl[1]+"_"+spl[2];
+            for( int a=3;a<spl.length;a++)
+                name+="_"+spl[a];
+            return new FlagVal().group(spl[0]).name(name);
+        }
         return new FlagVal().name(spl[0]);
     }